requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
httpie==2.5.0
idna==3.2
Pygments==2.10.0
PyJWT==1.7.1
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.7
